Avoid unnecessary casts of size_of() to size_t

+++ lib/compress/zstd_fast.rs
@@ -892,7 +892,7 @@ unsafe fn ZSTD_compressBlock_fast_dictMatchState_generic(
         (istart.offset_from(base) as core::ffi::c_long as size_t).wrapping_add(srcSize) as u32;
     if (*ms).prefetchCDictTables != 0 {
         let hashTableBytes = ((1 as core::ffi::c_int as size_t) << (*dictCParams).hashLog)
-            .wrapping_mul(::core::mem::size_of::<u32>() as size_t);
+            .wrapping_mul(::core::mem::size_of::<u32>());
         let _ptr = dictHashTable as *const core::ffi::c_char;
         let _size = hashTableBytes;
         let mut _pos: size_t = 0;
